backend_new/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pymongo==4.15.2
motor==3.7.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.9
python-dotenv==1.1.1
email-validator==2.3.0
